template配下のコンポーネントを細分化した
