+++ pages/outputs/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import axios from 'axios';
-import Layout from '@/components/layouts/Layout';
-import OutputsWrapper from '@/features/outputs/components/OutputsWrapper';
 import cookie from 'cookie';
+import Layout from '@/components/layouts/Layout';
+import OutputList from '@/features/outputs/components/OutputList';
 import { GetServerSideProps, NextPage } from 'next';
 import { getNextApiHeaders } from '@/common/functions/api';
 import { OutputsProps } from '@/features/outputs/types/output';
@@ -11,7 +11,7 @@ const Index: NextPage<OutputsProps> = ({ outputs }) => {
   return (
     <div>
       <Layout>
-        <OutputsWrapper outputs={outputs} />
+        <OutputList outputs={outputs} />
       </Layout>
     </div>
   );
